+++ src/components/StyleDrawer.js
@@ -33,17 +33,21 @@
 // Top Level Modules
 //----------------------------------------------------------------//
 import React from 'react'
+import { SketchPicker as ColorPicker } from 'react-color'
 
 //----------------------------------------------------------------//
 // Material-UI Core Components
 //----------------------------------------------------------------//
+import Button from '@material-ui/core/Button'
 import Drawer from '@material-ui/core/Drawer'
+import Grid from '@material-ui/core/Grid'
 import { makeStyles } from '@material-ui/core/styles'
+import Typography from '@material-ui/core/Typography'
 
 //----------------------------------------------------------------//
 // Custom Class Styling
 //----------------------------------------------------------------//
-const drawerWidth = 240
+const drawerWidth = 750
 
 const useStyles = makeStyles(theme => ({
   flex: {
@@ -82,20 +86,56 @@ const useStyles = makeStyles(theme => ({
 }))
 
 //----------------------------------------------------------------//
-// Analysis Tool Component
+// Style Drawer Component
 //----------------------------------------------------------------//
 /**
  * 
  * @param {*} props 
  */
-const LabelStyleDrawer = (props) => {
+const StyleDrawer = (props) => {
   const classes = useStyles()
 
-  const [index, setIndex] = React.useState('')
-  const [color, setColor] = React.useState('#ff0000')
+  const [mgrsGZColor, setMgrsGZColor] = React.useState('#ffa500')
+  const [mgrsLineColor, setMgrsLineColor] = React.useState('#ffffff')
+  const [garsCellColor, setGarsCellColor] = React.useState('#ffa500')
+  const [garsQuadColor, setGarsQuadColor] = React.useState('#800080')
+  const [garsKPColor, setGarsKPColor] = React.useState('#ffffff')
+  const [buildingLabelColor, setBuildingLabelColor] = React.useState('#ffff00')
 
   const container = props.window !== undefined ? () => window().document.body : undefined
 
+  const handleSubmit = () => {
+    const newStyles = {
+      marker: {
+        layer: 'styles',
+      },
+      mgrs: {
+        gridzoneColor: mgrsGZColor,
+        lineColor: mgrsLineColor,
+        lineBehavior: {
+          auto: true,
+          gridZone: false,
+          hundredKm: false,
+          tenKm: false,
+          oneKm: false,
+          hundredM: false,
+          tenM: false,
+          oneM: false,
+        }
+      },
+      gars: {
+        cellColor: garsCellColor,
+        quadrantColor: garsQuadColor,
+        keypadColor: garsKPColor,
+      },
+      buildingLabel: {
+        color: buildingLabelColor,
+      },
+    }
+
+    props.submit('edit', newStyles)
+    props.onClose()
+  }
 
   return (
     <nav
@@ -110,10 +150,185 @@ const LabelStyleDrawer = (props) => {
         classes={{ paper: classes.drawerPaper, }}
         ModalProps={{ keepMounted: true }}
       >
-        test
+        <Grid
+          container
+          direction='row'
+          justify='center'
+        >
+          <Typography
+            variant='body1'
+          >
+            MGRS Styles
+          </Typography>
+          <Grid
+            container
+            direction='row'
+            justify='center'
+          >
+            <Grid
+              item
+              xs={4}
+            >
+              <Typography
+                className={classes.marginsMd}
+                variant='body2'
+              >
+                Gridzone Color
+              </Typography>
+              <ColorPicker
+                className={classes.marginsMd}
+                color={mgrsGZColor}
+                disableAlpha={true}
+                onChange={color => setMgrsGZColor(color.hex)}
+              />
+            </Grid>
+            <Grid
+              item
+              xs={4}
+            >
+              <Typography
+                className={classes.marginsMd}
+                variant='body2'
+              >
+                Line Color
+              </Typography>
+              <ColorPicker
+                className={classes.marginsMd}
+                color={mgrsLineColor}
+                disableAlpha={true}
+                onChange={color => setMgrsLineColor(color.hex)}
+              />
+            </Grid>
+            <Grid
+              item
+              xs={4}
+            >
+              <Typography
+                className={classes.marginsMd}
+                variant='body2'
+              >
+                Line Behavior
+              </Typography>
+            </Grid>
+          </Grid>
+          <Typography
+            variant='body1'
+          >
+            GARS Styles
+          </Typography>
+          <Grid
+            container
+            direction='row'
+            justify='center'
+          >
+            <Grid
+              item
+              xs={4}
+            >
+              <Typography
+                className={classes.marginsMd}
+                variant='body2'
+              >
+                Cell Color
+              </Typography>
+              <ColorPicker
+                className={classes.marginsMd}
+                color={garsCellColor}
+                disableAlpha={true}
+                onChange={color => setGarsCellColor(color.hex)}
+              />
+            </Grid>
+            <Grid
+              item
+              xs={4}
+            >
+              <Typography
+                className={classes.marginsMd}
+                variant='body2'
+              >
+                Quadrant Color
+              </Typography>
+              <ColorPicker
+                className={classes.marginsMd}
+                color={garsQuadColor}
+                disableAlpha={true}
+                onChange={color => setGarsQuadColor(color.hex)}
+              />
+            </Grid>
+            <Grid
+              item
+              xs={4}
+            >
+              <Typography
+                className={classes.marginsMd}
+                variant='body2'
+              >
+                Keypad Color
+              </Typography>
+              <ColorPicker
+                className={classes.marginsMd}
+                color={garsKPColor}
+                disableAlpha={true}
+                onChange={color => setGarsKPColor(color.hex)}
+              />
+            </Grid>
+          </Grid>
+          <Typography
+            variant='body1'
+          >
+            Building Label Styles
+          </Typography>
+          <Grid
+            container
+            direction='row'
+            justify='center'
+          >
+            <Grid
+              item
+              xs={4}
+            >
+              <Typography
+                className={classes.marginsMd}
+                variant='body2'
+              >
+                Label Color
+              </Typography>
+              <ColorPicker
+                className={classes.marginsMd}
+                color={buildingLabelColor}
+                disableAlpha={true}
+                onChange={color => setBuildingLabelColor(color.hex)}
+              />
+            </Grid>
+            <Grid
+              item
+              xs={4}
+            >
+            </Grid>
+            <Grid
+              item
+              xs={4}
+            >
+            </Grid>
+          </Grid>
+        </Grid>
+        <Grid
+          container
+          direction='row'
+          justify='center'
+        >
+          <Button
+            className={classes.marginsMd}
+            color='primary'
+            onClick={() => handleSubmit()}
+            variant='contained'
+          >
+            Save Changes
+          </Button>
+        </Grid>
       </Drawer>
     </nav>
   )
 }
 
-export default LabelStyleDrawer
+export default StyleDrawer